% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ASRM.R
\name{get_asrm}
\alias{get_asrm}
\title{get_asrs function}
\usage{
get_asrm(dataset, completers = T)
}
\arguments{
\item{dataset}{original dataset "ASRM" from the bundle}

\item{completers}{boolean parameter, if True filters out participants that are not labeled as completers}
}
\value{
dataframe with 3 columns:
        PIN, asrm_sum, asrm_cat
}
\description{
get_asrs function
}
