% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ASRS.R
\name{get_asrs}
\alias{get_asrs}
\title{get_asrs function}
\usage{
get_asrs(dataset, subscales = F, completers = T)
}
\arguments{
\item{dataset}{original dataset "ASRS" from the bundle}

\item{completers}{boolean parameter, if True filters out participants that are not labeled as completers}
}
\value{
dataframe with 3 columns:
        PIN, response, asrs_cat
}
\description{
get_asrs function
}
