% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_AUDIT.R
\name{get_audit}
\alias{get_audit}
\title{get_audit function}
\usage{
get_audit(dataset, subscales = F, completers = T)
}
\arguments{
\item{dataset}{original dataset "AUDIT" from the bundle}

\item{subscales}{boolean parameter, if True includes to the returned dataframe moves subscales}

\item{completers}{boolean parameter, if True filters out participants that are not labeled as completers}
}
\value{
either dataframe with 4 columns:
        "PIN", "audit_sum", "audit_cat", "audit_sev" or dataframe with 7 columns: "PIN", "audit_sum", "audit_cat", "audit_sev", "audit_sym_ac", "audit_sym_bd", "audit_sym_probs"
}
\description{
get_audit function
}
