% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DASS.R
\name{get_dass}
\alias{get_dass}
\title{get_dass function}
\usage{
get_dass(dataset, subscales = F, completers = T)
}
\arguments{
\item{dataset}{original dataset "DASS" from the bundle}

\item{subscales}{boolean parameter, if True includes to the returned dataframe moves subscales}

\item{completers}{boolean parameter, if True filters out participants that are not labeled as completers}
}
\value{
either dataframe with 2 columns:
        PIN, dass_sum or dataframe with 8 columns: "PIN", "dass_sum","depression", "anxiety", "stress", "depression_cat", "anxiety_cat", "stress_cat"
}
\description{
get_dass function
}
