% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_demographics.R
\name{get_demographics}
\alias{get_demographics}
\title{get_demographics function}
\usage{
get_demographics(dataset, completers = T, convert = F, checks = T)
}
\arguments{
\item{dataset}{original dataset "demographics" from the bundle}

\item{completers}{boolean parameter, if True filters out participants that are not labeled as completers}

\item{convert}{boolean parameter, if True converts all categorical values into integers}

\item{checks}{boolean parameter, if True checks for improbable values in age, weight and height columns}
}
\value{
dataframe with eleven columns:
        PIN, gender, age,height,weight,BMI, marital status, education,employment, income, location
}
\description{
get_demographics function
}
