% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eating.R
\name{get_eating}
\alias{get_eating}
\title{get_eating function}
\usage{
get_eating(dataset, subscales = F, completers = T)
}
\arguments{
\item{dataset}{original dataset "EAT-26" from the bundle}

\item{subscales}{boolean parameter, if True includes to the returned dataframe eating subscales}

\item{completers}{boolean parameter, if True filters out participants that are not labeled as completers}
}
\value{
either dataframe with 3 columns:
        PIN, response, oci_cat or dataframe with 6 columns: PIN, eat_sum, eat_cat, eat_sym_diet, eat_sym_bul, eat_sym_oral
}
\description{
get_eating function
}
