% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GAD.R
\name{get_gad}
\alias{get_gad}
\title{get_gad function}
\usage{
get_gad(dataset, subscales = F, completers = T)
}
\arguments{
\item{dataset}{original dataset "GAD-7" from the bundle}

\item{subscales}{boolean parameter, if True includes to the returned dataframe moves subscales}

\item{completers}{boolean parameter, if True filters out participants that are not labeled as completers}
}
\value{
either dataframe with 5 columns:
        PIN, response, gad_sum, gad_cat, gad_sev or dataframe with 7 columns:PIN, response, gad_sum, gad_cat, gad_sev, gad_sym_nsom, gad_symp_som
}
\description{
get_gad function
}
