% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ISI.R
\name{get_isi}
\alias{get_isi}
\title{get_isi function}
\usage{
get_isi(dataset, completers = T)
}
\arguments{
\item{dataset}{original dataset "ASRM" from the bundle}

\item{completers}{boolean parameter, if True filters out participants that are not labeled as completers}
}
\value{
dataframe with 3 columns:
        PIN, isi_sum, isi_cat
}
\description{
get_isi function
}
