% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LSAS.R
\name{get_lsas}
\alias{get_lsas}
\title{get_lsas function}
\usage{
get_lsas(dataset, subscales = F, completers = T)
}
\arguments{
\item{dataset}{original dataset "LSAS" from the bundle}

\item{subscales}{boolean parameter, if True includes to the returned dataframe moves subscales}

\item{completers}{boolean parameter, if True filters out participants that are not labeled as completers}
}
\value{
either dataframe with 4 columns:
        PIN, lsas_sum, lsas_cat, lsas_sev or dataframe with 8 columns: "PIN", "lsas_sum", "lsas_cat", "lsas_sev", "lsas_sym_sif" "lsas_sym_sia" "lsas_sym_pf"  "lsas_sym_pa"
}
\description{
get_lsas function
}
