% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_OCI.R
\name{get_oci}
\alias{get_oci}
\title{get_OCI function}
\usage{
get_oci(dataset, subscales = F, completers = T)
}
\arguments{
\item{dataset}{original dataset "OCI" from the bundle}

\item{subscales}{boolean parameter, if True includes to the returned dataframe OCI subscales}

\item{completers}{boolean parameter, if True filters out participants that are not labeled as completers}
}
\value{
either dataframe with 3 columns:
        PIN, response, oci_cat or dataframe with 9 columns: PIN, response, oci_cat, sym_wash, sym_obsess,sym_hoard,sym_order,sym_check,sym_ment
}
\description{
get_OCI function
}
