% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pgsi.R
\name{get_pgsi}
\alias{get_pgsi}
\title{get_pgsi function}
\usage{
get_pgsi(dataset, subscales = F, completers = T)
}
\arguments{
\item{dataset}{original dataset "PGSI" from the bundle}

\item{completers}{boolean parameter, if True filters out participants that are not labeled as completers}
}
\value{
dataframe with 4 columns:
        PIN, pgsi_sum, pgsi_cat, pgsi_severity
}
\description{
get_pgsi function
}
