% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pid.R
\name{get_pid}
\alias{get_pid}
\title{get_pid function}
\usage{
get_pid(dataset, subscales = F, completers = T)
}
\arguments{
\item{dataset}{original dataset "PID-5-BF" from the bundle}

\item{subscales}{boolean parameter, if True includes to the returned dataframe moves subscales}

\item{completers}{boolean parameter, if True filters out participants that are not labeled as completers}
}
\value{
either dataframe with 3 columns:
        PIN, pid_sum, pid_cat or dataframe with 8 columns: "PIN", "pid_sum","pid_cat", "pid_sym_na", "pid_sym_det", "pid_sym_antag", "pid_sym_disin", "pid_sym_psych"
}
\description{
get_pid function
}
