% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prime.R
\name{get_prime}
\alias{get_prime}
\title{get_prime function}
\usage{
get_prime(dataset, subscales = F, completers = T)
}
\arguments{
\item{dataset}{original dataset "PRIME-R" from the bundle}

\item{completers}{boolean parameter, if True filters out participants that are not labeled as completers}
}
\value{
dataframe with 4 columns:
        PIN, cutoff_6, cutoff_5, prime_cat
}
\description{
get_prime function
}
