% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_RAADS.R
\name{get_raads}
\alias{get_raads}
\title{get_raads function}
\usage{
get_raads(dataset, subscales = F, completers = T)
}
\arguments{
\item{dataset}{original dataset "RAADS-14" from the bundle}

\item{subscales}{boolean parameter, if True includes to the returned dataframe moves subscales}

\item{completers}{boolean parameter, if True filters out participants that are not labeled as completers}
}
\value{
either dataframe with 3 columns:
        PIN, raads_sum, raads_cat or dataframe with 6 columns: "PIN", "raads_sum", "raads_cat", "raads_sym_md","raads_sym_sa","raads_sym_sr"
}
\description{
get_raads function
}
