% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SDS.R
\name{get_sds}
\alias{get_sds}
\title{get_sds function}
\usage{
get_sds(dataset, subscales = F, completers = T)
}
\arguments{
\item{dataset}{original dataset "SDS" from the bundle}

\item{completers}{boolean parameter, if True filters out participants that are not labeled as completers}
}
\value{
dataframe with 3 columns:
        PIN, sds_sum, sds_cat, sds_sev
}
\description{
get_sds function
}
