% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smoking.R
\name{get_smoking}
\alias{get_smoking}
\title{get_smoking function}
\usage{
get_smoking(dataset1, dataset2, subscales = F, completers = T)
}
\arguments{
\item{completers}{boolean parameter, if True filters out participants that are not labeled as completers}

\item{dataset}{original dataset "smoking_status" and "FTND" from the bundle}
}
\value{
dataframe with 4 columns:
        PIN, response, smoking_status, ftnd_status
}
\description{
get_smoking function
}
