% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yiat.R
\name{get_yiat}
\alias{get_yiat}
\title{get_yiat function}
\usage{
get_yiat(dataset, subscales = F, completers = T)
}
\arguments{
\item{dataset}{original dataset "YIAT" from the bundle}

\item{subscales}{boolean parameter, if True includes to the returned dataframe moves subscales}

\item{completers}{boolean parameter, if True filters out participants that are not labeled as completers}
}
\value{
either dataframe with 4 columns:
        PIN, phq_sum, phq_cat, phq_sev or dataframe with 6 columns: "PIN", "yiat_sum", "yiat_cat", "yiat_sev", "yiat_sym_loctm", "yiat_sym_csp"
}
\description{
get_yiat function
}
